requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==41.0.7
openai==1.3.0
google-generativeai==0.3.0
youtube-transcript-api==0.6.1
PyPDF2==3.0.1
pypdf==3.17.0
faiss-cpu>=1.9.0
numpy>=1.24.3,<1.27.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow>=10.0.0
moviepy==1.0.3
pydub==0.25.1

